Creating multiple book rows
