+++ frontend/src/pages/Library.js
@@ -1,10 +1,12 @@
 import React from "react";
 import BookTable from "../components/BookTable.js";
+import TableHeader from "../components/TableHeader.js";
 
 export default function Library() {
     return (
-        <div>
-           <h2>Library</h2>
-        </div>
+        <>
+            <TableHeader />
+            <BookTable />
+        </>
     );
 }
